extracted methods to manage basket items from app
